refactor(velvet): consolidate Hooks test fixtures
Merge single-seam fixture files into topical fixtures (6 clusters:
Memoize props-bail interactions, woven-behavior E2E, weaver
reflection probes, UseFrame dispatcher behavior, UseStore tearing
parity). Test bodies are preserved verbatim; colliding scaffolding
statics and component hosts are disambiguated with topical prefixes.
