--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/MemoGenericOverloadTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 904d31215c0a94a71a439815ea335ba3
