--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/MemoizeErrorBoundaryComboTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 005712a782655480b91e369c08e30e5e
