--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/MemoizeFunctionalE2ETests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 648e51eb8edcb4500bf9efc716caf34c
