--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/MemoizeSuspenseInteractionTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: d3d0824cd931a4e7f9a6aa7f0230bbce
