--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/MemoizeWovenTupleBindingTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 75743a45c22f4478c81f9c1998553cc6
