--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/OpenDispatchHookSafetyTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 66c51bc16fad4dce84687d5df466136f
