--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/UseFrameDispatcherReentrancyTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0854f792894da40b38593c9a11614a59
