--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/UseFrameKeyedReorderTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 4ac54f2722ccc4908a718dac2beb112c
