--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/UseFramePerFrameContractTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f89125f56e631407095c5c6eca71d565
