--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/UseFramePriorityTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: cec42802c61d24d49b0befd6d78a127b
