--- Packages/com.velvet.core/Runtime/Hooks/Tests/Editor/UseStoreTearingParityTests.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 72c72fd7607842b2bfe36f045157e094
